+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 * Add mute/unmute events support
 * Allow to build history transcript inside SDK
+* Add parameters to triggers
 
 ## 2022-04-20 v0.9.7